@@ -74,7 +58,7 @@
     "commitizen": "4.3.1",
     "concurrently": "9.0.1",
     "cz-conventional-changelog": "3.3.0",
-    "eslint": "8.57.1",
+    "eslint": "9.12.0",
     "github-markdown-css": "5.7.0",
     "husky": "9.1.6",
     "jest-environment-jsdom": "29.7.0",
