@@ -1,6 +1,16 @@
 import js from '@eslint/js'
 
-const eslintNoUnusedVars = { rules: { 'no-unused-vars': 'off' } }
-const eslintSortKeys = { rules: { 'sort-keys': ['error', 'asc', { caseSensitive: true, natural: true }] } }
-
-export const eslintConfigs = [js.configs.recommended, eslintNoUnusedVars, eslintSortKeys]
+export const eslintConfigs = [
+  {
+    name: 'eslint/recommended',
+    ...js.configs.recommended,
+  },
+  {
+    name: 'eslint/no-unused-vars',
+    rules: { 'no-unused-vars': 'off' },
+  },
+  {
+    name: 'eslint/sort-keys',
+    rules: { 'sort-keys': ['error', 'asc', { caseSensitive: true, natural: true }] },
+  },
+]
