@@ -2,8 +2,13 @@ import jest from 'eslint-plugin-jest'
 
 export const jestConfigs = [
   {
-    files: ['**/__tests__/**/*.+(j|t)s?(x)', '**/?(*.)+(spec|test).+(j|t)s?(x)'],
+    files: ['**/*.test.@(j|t)s?(x)'],
+    name: 'jest',
     ...jest.configs['flat/all'],
   },
-  { rules: { 'jest/max-expects': ['error', { max: 10 }] } },
+  {
+    files: ['**/*.test.@(j|t)s?(x)'],
+    name: 'jest/max-expects',
+    rules: { 'jest/max-expects': ['error', { max: 10 }] },
+  },
 ]
