@@ -1,6 +1,10 @@
 import perfectionist from 'eslint-plugin-perfectionist'
 
 export const perfectionistConfigs = [
-  perfectionist.configs['recommended-natural-legacy'],
-  { settings: { perfectionist: { ignoreCase: false } } },
+  {
+    name: 'perfectionist',
+    ...perfectionist.configs['recommended-natural-legacy'],
+    plugins: { perfectionist },
+    settings: { perfectionist: { ignoreCase: false } },
+  },
 ]
