@@ -1,3 +1,8 @@
 import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended'
 
-export const prettierConfigs = [eslintPluginPrettierRecommended]
+export const prettierConfigs = [
+  {
+    name: 'prettier',
+    ...eslintPluginPrettierRecommended,
+  },
+]
