@@ -1,8 +1,14 @@
+import { fixupPluginRules } from '@eslint/compat'
 import testingLibrary from 'eslint-plugin-testing-library'
 
 export const testingLibraryConfigs = [
   {
-    files: ['**/__tests__/**/*.+(j|t)s?(x)', '**/?(*.)+(spec|test).+(j|t)s?(x)'],
+    files: ['**/*.test.@(j|t)s?(x)'],
     ...testingLibrary.configs['flat/react'],
+    plugins: { 'testing-library': fixupPluginRules(testingLibrary) },
+  },
+  {
+    name: 'testing-library/render-result-naming-convention',
+    rules: { 'testing-library/render-result-naming-convention': 'off' },
   },
 ]
