@@ -1,4 +1,4 @@
-export * from './presets-v9/hybrid'
-export * from './presets-v9/javascript'
-export * from './presets-v9/markdown'
-export * from './presets-v9/typescript'
+export * from './presets-v9/hybrid.js'
+export * from './presets-v9/javascript.js'
+export * from './presets-v9/markdown.js'
+export * from './presets-v9/typescript.js'
