@@ -1,3 +1,3 @@
-import { markdown } from './markdown'
+import { markdown } from './markdown.js'
 
 export const hybrid = markdown
