@@ -1,4 +1,4 @@
-import { mdxConfigs } from '../configs-v9/mdx'
-import { typescript } from './typescript'
+import { mdxConfigs } from '../configs-v9/mdx.js'
+import { typescript } from './typescript.js'
 
 export const markdown = [...typescript, ...mdxConfigs]
