@@ -1,4 +1,4 @@
-import { typescriptEslintConfigs } from '../configs-v9/typescript'
-import { javascript } from './javascript'
+import { typescriptEslintConfigs } from '../configs-v9/typescript.js'
+import { javascript } from './javascript.js'
 
 export const typescript = [...javascript, ...typescriptEslintConfigs]
